refactor: 抽出抽出 data 資料
