+++ src/features/ble/data/PairedApp.ts
@@ -1,0 +1,4 @@
+export interface PairedApp {
+  appId: string;
+  deviceName: string;
+}
