+++ src/features/ble/data/RegisterInfo.ts
@@ -1,0 +1,5 @@
+export interface RegisterInfo {
+  deviceName: string;
+  password: string;
+  appId: string;
+}
